Separate main executable as separate file from node logic
